+++ src/orquestra/sdk/_base/_qe/_qe_runtime.py
@@ -22,7 +22,7 @@ import pydantic
 import requests
 
 from orquestra.sdk import exceptions
-from orquestra.sdk._base import serde
+from orquestra.sdk._base import _dates, serde
 from orquestra.sdk._base._conversions._yaml_exporter import (
     pydantic_to_yaml,
     workflow_to_yaml,
@@ -885,7 +885,7 @@ class QERuntime(RuntimeInterface):
                 "Filtering by workspace or project is not supported on QE runtimes."
             )
 
-        now = datetime.now(timezone.utc)
+        now = _dates.now()
 
         # Grab the workflows we know about from the DB
         with WorkflowDB.open_project_db(self._project_dir) as db: