+++ tests/cli/dorq/ui/test_presenters.py
@@ -11,6 +11,7 @@ import pytest
 
 from orquestra import sdk
 from orquestra.sdk._base import serde
+from orquestra.sdk._base._dates import Instant
 from orquestra.sdk._base._spaces._structs import Project, Workspace
 from orquestra.sdk._base.cli._dorq._ui import _errors
 from orquestra.sdk._base.cli._dorq._ui import _models as ui_models
@@ -392,26 +393,30 @@ class TestConfigPresenter:
 DATA_DIR = Path(__file__).parent / "data"
 
 
-UTC_INSTANT = datetime(2023, 2, 24, 12, 26, 7, 704015, tzinfo=timezone.utc)
-ET_INSTANT_1 = datetime(
-    2023,
-    2,
-    24,
-    7,
-    26,
-    7,
-    704015,
-    tzinfo=timezone.utc,
+UTC_INSTANT = Instant(datetime(2023, 2, 24, 12, 26, 7, 704015, tzinfo=timezone.utc))
+ET_INSTANT_1 = Instant(
+    datetime(
+        2023,
+        2,
+        24,
+        7,
+        26,
+        7,
+        704015,
+        tzinfo=timezone.utc,
+    )
 )
-ET_INSTANT_2 = datetime(
-    2023,
-    2,
-    24,
-    7,
-    28,
-    37,
-    123,
-    tzinfo=timezone.utc,
+ET_INSTANT_2 = Instant(
+    datetime(
+        2023,
+        2,
+        24,
+        7,
+        28,
+        37,
+        123,
+        tzinfo=timezone.utc,
+    )
 )
 
 
@@ -494,7 +499,7 @@ class TestWorkflowRunPresenter:
 
 class TestPromptPresenter:
     class TestWorkspacesList:
-        def test_workspaces_list_to_prompt_no_studio(self):
+        def test_workspaces_list_to_prompt(self):
             # given
             workspace1 = Workspace("id1", "name1")
             workspace2 = Workspace("id2", "name2")
@@ -515,7 +520,7 @@ class TestPromptPresenter:
             assert "id3" in labels[2]
             assert "name3" in labels[2]
 
-        def test_workspace_list_to_prompt_studio_auto(self, monkeypatch):
+        def test_workspace_list_to_prompt_env_var_set(self, monkeypatch):
             current_workspace_id = "id3"
             monkeypatch.setenv("ORQ_CURRENT_WORKSPACE", current_workspace_id)
             workspace1 = Workspace("id1", "name1")
@@ -527,7 +532,7 @@ class TestPromptPresenter:
 
             # when
             labels, workspaces = presenter.workspaces_list_to_prompt(
-                workspaces=workspace_list, config_name="auto"
+                workspaces=workspace_list
             )
 
             # current workspace should be the 1st one in the list
@@ -542,31 +547,8 @@ class TestPromptPresenter:
             assert "id2" in labels[2]
             assert "name2" in labels[2]
 
-        def test_workspace_list_to_prompt_studio_not_auto(self, monkeypatch):
-            current_workspace_id = "id3"
-            monkeypatch.setenv("ORQ_CURRENT_WORKSPACE", current_workspace_id)
-            workspace1 = Workspace("id1", "name1")
-            workspace2 = Workspace("id2", "name2")
-            workspace3 = Workspace(current_workspace_id, "name3")
-
-            workspace_list = [workspace1, workspace2, workspace3]
-            presenter = _presenters.PromptPresenter()
-
-            # when
-            labels, workspaces = presenter.workspaces_list_to_prompt(
-                workspaces=workspace_list, config_name="different_config"
-            )
-
-            assert workspaces == workspace_list
-            assert "id1" in labels[0]
-            assert "name1" in labels[0]
-            assert "id2" in labels[1]
-            assert "name2" in labels[1]
-            assert "id3" in labels[2]
-            assert "name3" in labels[2]
-
     class TestProjectList:
-        def test_project_list_to_prompt_no_studio(self):
+        def test_project_list_to_prompt(self):
             # given
             project1 = Project("id1", "ws", "name1")
             project2 = Project("id2", "ws", "name2")
@@ -585,7 +567,7 @@ class TestPromptPresenter:
             assert "id3" in labels[2]
             assert "name3" in labels[2]
 
-        def test_workspace_list_to_prompt_studio_auto(self, monkeypatch):
+        def test_workspace_list_to_prompt_env_set(self, monkeypatch):
             current_project_id = "id3"
             monkeypatch.setenv("ORQ_CURRENT_PROJECT", current_project_id)
             project1 = Project("id1", "ws", "name1")
@@ -596,9 +578,7 @@ class TestPromptPresenter:
             presenter = _presenters.PromptPresenter()
 
             # when
-            labels, projects = presenter.project_list_to_prompt(
-                project_list, config_name="auto"
-            )
+            labels, projects = presenter.project_list_to_prompt(project_list)
 
             # current project should be the 1st one in the list
             assert projects == [project3, project1, project2]
@@ -611,26 +591,3 @@ class TestPromptPresenter:
 
             assert "id2" in labels[2]
             assert "name2" in labels[2]
-
-        def test_workspace_list_to_prompt_studio_not_auto(self, monkeypatch):
-            current_project_id = "id3"
-            monkeypatch.setenv("ORQ_CURRENT_PROJECT", current_project_id)
-            project1 = Project("id1", "ws", "name1")
-            project2 = Project("id2", "ws", "name2")
-            project3 = Project(current_project_id, "ws", "name3")
-
-            project_list = [project1, project2, project3]
-            presenter = _presenters.PromptPresenter()
-
-            # when
-            labels, projects = presenter.project_list_to_prompt(
-                project_list, config_name="different_config"
-            )
-
-            assert projects == project_list
-            assert "id1" in labels[0]
-            assert "name1" in labels[0]
-            assert "id2" in labels[1]
-            assert "name2" in labels[1]
-            assert "id3" in labels[2]
-            assert "name3" in labels[2]