+++ tests/sdk/v2/test_in_process_runtime.py
@@ -8,14 +8,14 @@ functionality is covered inside tests/v2/test_api.py.
 We need this file because test_api.py might be too coarse for some scenarios.
 When adding new tests, please consider that suite first.
 """
-from datetime import datetime, timedelta, timezone
+from datetime import timedelta
 from unittest.mock import create_autospec
 
 import pytest
 
 from orquestra import sdk
 from orquestra.sdk import exceptions
-from orquestra.sdk._base import serde
+from orquestra.sdk._base import _dates, serde
 from orquestra.sdk._base._in_process_runtime import InProcessRuntime
 from orquestra.sdk._base._spaces._structs import ProjectRef
 from orquestra.sdk._base._testing._example_wfs import (
@@ -234,9 +234,7 @@ class TestListWorkflowRuns:
         # Given
         run_id = runtime.create_workflow_run(wf_def, None)
         _ = runtime.create_workflow_run(wf_def, None)
-        runtime._start_time_store[run_id] = datetime.now(timezone.utc) - timedelta(
-            days=1
-        )
+        runtime._start_time_store[run_id] = _dates.now() - timedelta(days=1)
 
         # When
         wf_runs = runtime.list_workflow_runs(max_age=timedelta(minutes=1))